Day 17/trip_planner/requirements.txt
langgraph
python-dotenv
certifi
fastapi
uvicorn
jinja2
python-multipart
streamlit
reportlab
chromadb
langchain-openai
